[#21] Add UI Test
