+++ shared/src/commonMain/kotlin/co/nimblehq/avishek/phong/kmmic/helper/ios/UITestLogout.kt
@@ -1,0 +1,14 @@
+package co.nimblehq.avishek.phong.kmmic.helper.ios
+
+import co.nimblehq.avishek.phong.kmmic.data.local.datasource.TokenLocalDataSource
+import org.koin.core.component.KoinComponent
+import org.koin.core.component.inject
+
+//object UITestLogout: KoinComponent {
+//
+//    private val localDataSource: TokenLocalDataSource by inject()
+//
+//    fun removeAccessToken() {
+//        localDataSource.removeToken()
+//    }
+//}
