--- shared/src/iosMain/kotlin/co/nimblehq/avishek/phong/kmmic/FlowAdapter.kt
@@ -1,35 +1,0 @@
-package co.nimblehq.avishek.phong.kmmic
-
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.onEach
-import kotlinx.coroutines.flow.catch
-import kotlinx.coroutines.flow.onCompletion
-import kotlinx.coroutines.flow.launchIn
-
-class FlowAdapter<T : Any>(
-    private val scope: CoroutineScope,
-    private val flow: Flow<T>
-) {
-    fun subscribe(
-        onEach: (item: T) -> Unit,
-        onComplete: () -> Unit,
-        onThrow: (error: Throwable) -> Unit
-    ): Canceller = JobCanceller(
-        flow.onEach { onEach(it) }
-            .catch { onThrow(it) }
-            .onCompletion { onComplete() }
-            .launchIn(scope)
-    )
-}
-
-interface Canceller {
-    fun cancel()
-}
-
-private class JobCanceller(private val job: Job) : Canceller {
-    override fun cancel() {
-        job.cancel()
-    }
-}
